3.0.5
Additional unit testing. Output should go to the console as a string so that critical child elements are not [object] rather than being fully serialized.
Fixed minor issue with multiple errors being reported in one statement.
Updated debut and trace to output to debug instead of info, to make it more visible in situations where console log type is referenced.